move and rename prometheus_exporter.py
